+++ BitmovinComscoreAnalytics/Classes/ComScoreStreamingAnalytics.swift
@@ -9,7 +9,7 @@ import Foundation
 import BitmovinPlayer
 
 public class ComScoreStreamingAnalytics {
-    let comScoreAdapter: ComScoreBitmovinAdapter
+    let adapter: ComScoreBitmovinAdapter
 
     // MARK: - initializer
     /**
@@ -20,25 +20,29 @@ public class ComScoreStreamingAnalytics {
      - metadata: ComScore metadata associated with the content you are tracking
      */
     init(bitmovinPlayer: Player, configuration: ComScoreConfiguration, metadata: ComScoreMetadata) {
-        self.comScoreAdapter = ComScoreBitmovinAdapter(player: bitmovinPlayer, configuration: configuration, metadata: metadata)
+        adapter = ComScoreBitmovinAdapter(
+            player: bitmovinPlayer,
+            configuration: configuration,
+            metadata: metadata
+        )
     }
 
     deinit {
-        self.comScoreAdapter.destroy()
+        adapter.destroy()
     }
 
     /**
      Destroy ComScoreStreamingAnalytics and unregister it from player
      */
     public func destroy() {
-        self.comScoreAdapter.destroy()
+        adapter.destroy()
     }
 
     /**
      Update metadata for tracked source. This should be called when changing sources.
      */
     public func update(metadata: ComScoreMetadata) {
-        comScoreAdapter.update(metadata: metadata)
+        adapter.update(metadata: metadata)
     }
 
     /**
@@ -48,7 +52,7 @@ public class ComScoreStreamingAnalytics {
      - value: The label value
      */
     public func setPersistentLabel(label: String, value: String) {
-        comScoreAdapter.setPersistentLabel(label: label, value: value)
+        adapter.setPersistentLabel(label: label, value: value)
     }
 
     /**
@@ -57,6 +61,20 @@ public class ComScoreStreamingAnalytics {
      - label: The labels to set
      */
     public func setPersistentLabels(labels: [String: String]) {
-        comScoreAdapter.setPersistentLabels(labels: labels)
+        adapter.setPersistentLabels(labels: labels)
+    }
+    
+    /**
+     Stop ComScore streaming analytics tracking
+     */
+    public func stopTracking() {
+        adapter.stop()
+    }
+    
+    /**
+     Start ComScore streaming analytics tracking
+     */
+    public func startTracking() {
+        adapter.resume()
     }
 }