+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.7.1]
+
+### Added
+- Stop ComScore streaming analytics tracking API
+- Start ComScore streaming analytics tracking API
+
 ## [1.7.0]
 
 ### Changed